\name{MethylCNVDataSet-class}
\Rdversion{1.0}
\docType{class}
\alias{MethylCNVDataSet}
\title{Class \code{"MethylCNVDataSet"}}
\description{
  This class holds Illumina 450k methylation microarray data and annotations for CNV calling.
}
\se
  %MethylCNVDataSetFromRGChannelSet("")
}
\section{Methods}
\describe{
  \item{sampleNames}{\code{signature(object = "CNVObject")}: ...}
}
\author{

}
\examples{

}
\keyword{classes}